requirements.txt
streamlit>=1.40.0
pdfplumber>=0.11.0
pandas>=2.2.0
duckdb>=1.1.0
plotly>=5.24.0
sqlparse>=0.5.0
python-dotenv>=1.0.0
tabulate>=0.9.0   
pytest>=8.3.0
openai>=1.50.0